Split compile() to bytecode_size() and compile()
bytecode_size() simply returns the number of bytes that the given proc
node will compile into. compile() will then take a buffer and do all
the compiling. Because of this, compile() no longer needs an external
definition of has_errors().

On an aside, once multiple processors are compiled at the same time,
this makes is possible to compile all bytecode into a single
buffer. Depending how I do it, this could be either very convenient or
very dangerous!