Simplify code, use more graceful cascading, etc
This might be better split into two commits, but since they're both a
part of the same yak-shaving session, I think it's a good fit.

The first part is that instead of manually compiling a uint16 value
into octets, to let the compiler decide how to pack it via
casting. The second part is removing edge conditions by allowing
code (specifically involved with memory allocation) to cascade.