Remove stack nodes from the specification
Stack nodes as it was written only had one port, %elm. Following the
rule of each port only being able to attach to one node, someone would
only be able to either push items to the stack node, or pop them, but
not both. This would make the stack node useless.